backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "langchain-anthropic>=1.3.4",
    "langchain-core>=1.2.16",
    "langgraph>=0.1",
    "lime>=0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "pydantic>=2",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.4",
    "shap>=0.45",
    "python-dotenv>=1.0",
    "uvicorn[standard]>=0.41.0",
]

[project.optional-dependencies]
gemini = [
    "langchain-google-genai>=4.2.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
package = false
